Conflicts:
	computer/src/uk/ac/ed/inf/sdp/group4/strategy/Match.java

+++ computer/src/uk/ac/ed/inf/sdp/group4/domain/InvalidAngleException.java
@@ -2,7 +2,7 @@ package uk.ac.ed.inf.sdp.group4.domain;
 
 public class InvalidAngleException extends Exception
 {
-	public InvalidAngleException(int angle)
+	public InvalidAngleException(double angle)
 	{
 		super("Angle was not in range 0-359 (" + angle + ")");
 	}
